Move util files to a new directory
