+++ src/__tests__/visualizer/algorithm/sortingalgorithms/bubbleSort.test.js
@@ -1,5 +1,5 @@
 import bubbleSort from '../../../../visualizer/algorithm/sortingalgorithms/bubbleSort';
-import { assertSortSuccess } from '../testutil/TestUtil';
+import { assertSortSuccess } from '../../../../_testutil/TestUtil';
 import {
   ARRAY_DUPLICATES,
   ARRAY_EXPECTED_RANDOM_MIXED,
@@ -12,7 +12,7 @@ import {
   ARRAY_STRICTLY_ASCENDING,
   ARRAY_STRICTLY_DESCENDING,
   EMPTY_ARRAY,
-} from '../testutil/ArraysUtil';
+} from '../../../../_testutil/ArraysUtil';
 
 test('Empty Array', () => {
   assertBubbleSortSuccess(EMPTY_ARRAY, EMPTY_ARRAY, EMPTY_ARRAY);
