+++ src/__tests__/visualizer/algorithm/sortingalgorithms/countingSort.test.js
@@ -1,5 +1,5 @@
 import countingSort from '../../../../visualizer/algorithm/sortingalgorithms/countingSort';
-import { assertSortSuccess } from '../testutil/TestUtil';
+import { assertSortSuccess } from '../../../../_testutil/TestUtil';
 import {
   ARRAY_DUPLICATES,
   ARRAY_EXPECTED_RANDOM_POSITIVE,
@@ -8,7 +8,7 @@ import {
   ARRAY_STRICTLY_ASCENDING,
   ARRAY_STRICTLY_DESCENDING,
   EMPTY_ARRAY,
-} from '../testutil/ArraysUtil';
+} from '../../../../_testutil/ArraysUtil';
 
 test('Empty Array', () => {
   assertCountSortSuccess(EMPTY_ARRAY, EMPTY_ARRAY);
